% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmr_filter}
\alias{cmr_filter}
\title{Tagged bird example}
\format{
A data frame with 11 rows and 8 variables:
\describe{
  \item{\code{ring_id}}{character ring number}
  \item{\code{date}}{double d}ate
  \item{\code{time}}{double time}
  \item{\code{sex}}{character sex}
  \item{\code{tars_mm}}{double tarsus}
  \item{\code{weight_g}}{double weight in gram}
  \item{\code{blood_infection}}{character tested for blood infection}
  \item{\code{infection_type}}{character infection type}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
cmr_filter
}
\description{
This dataset is one example bird to show how the package works
}
\keyword{datasets}
