% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re.R
\name{re}
\alias{re}
\title{recapture info}
\usage{
re(ring)
}
\arguments{
\item{ring}{the ring numer}
}
\value{
a dataframe with 1 more observation & prints information about previous captures
}
\description{
recapture info
}
\examples{
\dontrun{
re(123)
 }
}
