% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re.R
\name{re}
\alias{re}
\title{recapture}
\usage{
re(ring)
}
\arguments{
\item{ring}{ring_id}
}
\value{
dataframe with +1 observation
}
\description{
helps with recaptures
}
\details{
none
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{bind}}
}
